requirements.txt
mingus
pylint
qdarkgraystyle
pyqt5
pyqt5-tools
numpy
pandas
scipy
pyaudio
pyttsx3
speechRecognition
wikipedia
pafy
pygame
bs4
python-levenshtein
mutagen
selenium
lxml
playsound
pydub
youtube_search
youtube-dl